migrate all to fastapi based uvicorn app
