+++ mldock/templates/aws/src/assets.py
