+++ mldock/templates/generic/src/assets.py
