+++ tests/fixtures/containers/aws/src/assets.py
