--- tests/fixtures/containers/gcpv2/src/container/executor.sh
@@ -1,8 +1,0 @@
-#!/usr/bin/env bash
-script_name="$1"
-shift
-if [ "$script_name" = "train" ]; then
-    python ./src/container/training/train.py $@
-else
-    python ./src/container/prediction/serve.py
-fi
