--- tests/fixtures/containers/gcpv2/src/requirements.txt
@@ -1,4 +1,0 @@
-pandas
-scikit-learn
-numpy
-protobuf>=3.1
