+++ tests/fixtures/containers/generic/src/assets.py
